requirements.txt
paho-mqtt
pymongo
keyboard
pynput